move some files around, speed up gas price revalidation
