+++ modules/app/components/layout/Footer/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ShortFooter from './ShortFooter';
-// import LongFooter from './LongFooter';
-import LandingFooter from './LandingFooter';
+import LandingFooter from 'modules/home/components/LandingFooter';
 
 export default function Footer({ shorten = false }: { shorten: boolean }): React.ReactElement {
   if (shorten) {
